% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_drg_payments.R
\name{calculate_drg_stats}
\alias{calculate_drg_stats}
\title{Title}
\usage{
calculate_drg_stats(data, statistic = "mean")
}
\arguments{
\item{data}{dd}

\item{statistic}{dd}
}
\value{
xx
}
\description{
Title
}
\examples{
xx
}
